refactor: move Item,ItemRepository,Category from order to item package
