+++ backend/src/main/java/kr/co/programmers/cafe/domain/item/api/ItemController.java
@@ -21,6 +21,13 @@ import org.springframework.web.bind.annotation.RestController;
 public class ItemController {
     private final ItemService itemService;
 
+    /**
+     * 페이지네이션이 적용된 아이템 목록을 조회합니다.
+     *
+     * @param page 조회할 페이지 번호 (기본값: 0)
+     * @param size 페이지당 아이템 수 (기본값: 10)
+     * @return 페이지네이션이 적용된 아이템 목록
+     */
     @GetMapping
     public ResponseEntity<Page<ItemResponse>> readItems(
             @RequestParam(defaultValue = "0") int page,
@@ -30,12 +37,26 @@ public class ItemController {
         return ResponseEntity.ok(itemResponses);
     }
 
+    /**
+     * 특정 아이템의 상세 정보를 조회합니다.
+     *
+     * @param itemId 조회할 아이템의 ID
+     * @return 아이템 상세 정보
+     */
     @GetMapping("/{itemId}")
     public ResponseEntity<ItemResponse> readItem(@PathVariable Long itemId) {
         ItemResponse itemResponse = itemService.findById(itemId);
         return ResponseEntity.ok(itemResponse);
     }
 
+    /**
+     * 특정 아이템의 이미지를 조회합니다.
+     * 이미지 파일 확장자에 따라 적절한 MediaType을 설정합니다.
+     * 지원하는 이미지 형식: JPG, JPEG, GIF, PNG
+     *
+     * @param itemId 이미지를 조회할 아이템의 ID
+     * @return 이미지 리소스
+     */
     @GetMapping("/{itemId}/images")
     public ResponseEntity<Resource> readItemImages(@PathVariable Long itemId) {
         Resource resource = itemService.getImage(itemId);
