+++ backend/src/main/java/kr/co/programmers/cafe/domain/item/dao/ItemRepository.java
@@ -1,6 +1,6 @@
-package kr.co.programmers.cafe.domain.order.dao;
+package kr.co.programmers.cafe.domain.item.dao;
 
-import kr.co.programmers.cafe.domain.order.entity.Item;
+import kr.co.programmers.cafe.domain.item.entity.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ItemRepository extends JpaRepository<Item, Long> {
