+++ backend/src/main/java/kr/co/programmers/cafe/domain/item/entity/Category.java
@@ -1,4 +1,4 @@
-package kr.co.programmers.cafe.domain.order.entity;
+package kr.co.programmers.cafe.domain.item.entity;
 
 // 임시 카테고리
 public enum Category {
