+++ backend/src/main/java/kr/co/programmers/cafe/domain/order/entity/OrderItem.java
@@ -1,6 +1,14 @@
 package kr.co.programmers.cafe.domain.order.entity;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import kr.co.programmers.cafe.domain.item.entity.Item;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
