+++ backend/src/test/java/kr/co/programmers/cafe/domain/admin/AdminControllerTest.java
@@ -4,7 +4,7 @@ import kr.co.programmers.cafe.domain.item.app.ItemService;
 import kr.co.programmers.cafe.domain.item.dto.ItemCreateForm;
 import kr.co.programmers.cafe.domain.item.dto.ItemEditForm;
 import kr.co.programmers.cafe.domain.item.dto.ItemResponse;
-import kr.co.programmers.cafe.domain.order.entity.Category;
+import kr.co.programmers.cafe.domain.item.entity.Category;
 import kr.co.programmers.cafe.global.exception.ItemNotFoundException;
 import kr.co.programmers.cafe.global.util.FileManager;
 import org.junit.jupiter.api.BeforeEach;
