+++ backend/src/test/java/kr/co/programmers/cafe/domain/order/app/OrderServiceTest.java
@@ -1,23 +1,14 @@
 package kr.co.programmers.cafe.domain.order.app;
 
-import kr.co.programmers.cafe.domain.order.dao.ItemRepository;
-import kr.co.programmers.cafe.domain.order.dto.OrderItemRequest;
-import kr.co.programmers.cafe.domain.order.dto.OrderRequest;
-import kr.co.programmers.cafe.domain.order.entity.Category;
-import kr.co.programmers.cafe.domain.order.entity.Item;
+import kr.co.programmers.cafe.domain.item.dao.ItemRepository;
+import kr.co.programmers.cafe.domain.item.entity.Category;
+import kr.co.programmers.cafe.domain.item.entity.Item;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import java.util.List;
